requirements.txt
neo4j==5.28.1
openai==1.88.0
PyYAML==6.0.2
orjson==3.10.18
pandas==2.2.2
pypdf==5.6.0
typing_extensions==4.12.2
python-dotenv==1.0.1
pydantic==2.10.6
pydantic-settings==2.9.1
pydantic_core==2.27.2
chainlit==2.6.2
chromadb==1.0.16
rank_bm25==0.2.2
langfuse==3.2.4

langchain==0.3.26
langchain_community==0.3.27
langchain-neo4j==0.5.0
langchain-core==0.3.74
langchain_tavily==0.2.11
langchain-openai==0.3.28
langchain_chroma==0.2.5
langchain-postgres==0.0.13
langgraph-checkpoint==2.1.0
langgraph==0.6.4
langgraph-checkpoint-sqlite==2.0.11

neo4j_graphrag==1.9.1
rapidfuzz==3.13.0
fsspec>=2024.9.0,<2025.0.0
json-repair==0.44.1
pypdf>=5.1.0,<6.0.0
types-pyyaml>=6.0.12.20240917,<7.0.0.0